requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
cbor==1.0.0
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
colorlog==6.9.0
cryptography==45.0.5
cssselect==1.3.0
cssutils==2.11.1
dataclasses-json==0.6.7
datasets==4.0.0
dill==0.3.8
distro==1.9.0
docling @ git+https://github.com/docling-project/docling@c5f49dc2db43d38997196b7c49f90ea3b00c07ca
docling-core @ git+https://github.com/docling-project/docling-core@8433d285f6c6ee241d361614109bdff3634b62a5
docling-ibm-models==3.9.0
docling-parse==4.1.0
docx2txt==0.9
easyocr==1.7.2
einops==0.8.1
et-xmlfile==2.0.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
flagembedding==1.3.5
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
google-auth==2.40.3
gputil==1.4.0
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
ijson==3.4.0
imageio==2.37.0
imagesize==1.4.1
inscriptis==2.6.0
ir-datasets==0.5.11
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.12
latex2mathml==3.78.0
lazy-loader==0.4
lxml==5.4.0
lz4==4.4.4
markdown-it-py==3.0.0
marko==2.1.4
markupsafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
more-itertools==10.7.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
mypy-extensions==1.1.0
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnxruntime==1.22.1
openai==1.99.1
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.11.1
packaging==25.0
paddleocr==3.1.0
paddlex==3.1.3
pandas==2.3.1
pdfminer-six==20250506
pdfplumber==0.11.7
peft==0.17.0
pillow==11.3.0
pluggy==1.6.0
premailer==3.10.0
prettytable==3.16.0
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pylatexenc==2.10
pymupdf==1.26.3
pymupdf4llm==0.0.27
pypdfium2==4.30.0
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-magic==0.4.27
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pyyaml==6.0.2
rapidocr-onnxruntime==1.4.4
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
rtree==1.4.0
ruamel-yaml==0.18.14
ruamel-yaml-clib==0.2.12
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
semchunk==2.2.2
sentence-transformers==5.0.0
sentencepiece==0.2.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.42
starlette==0.47.2
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.9.0
tokenizers==0.21.4
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.54.1
trec-car-tools==2.6
triton==3.3.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
ujson==5.10.0
unlzw3==0.2.3
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
xlrd==2.0.2
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zlib-state==0.1.9
zstandard==0.23.0
